Update ES.yaml
ulm_heat: "Caliente" is a name and means "hot" in english. 
ulm_cool: "Enfriar"  is a verb and means "to cool" in english. In spanish, different words are used depending if it's a verb or an adjective.

I propose to change to "Calor" and "frio" as these are adjectives and are the usual name seen in climate devices