Update card_battery.yaml
For me the mdi:battery-charging icon was not working for iOS/iPad OS. The state of the battery_state sensor is "Charging". But a close look at the code and far as i understand it have a problem i think. If the variables.ulm_card_battery_charger_type_entity_id is not defined by the user or the state is not "null", "ac" or "wireless" the battery-charging icon will never be used. 

The strange thing is that the code below (far as i understand...) will pass the battery-charging icon if the state of variables.ulm_card_battery_battery_state_entity_id is not equal to "null" and the state of variables.ulm_card_battery_battery_state_entity_id is "charging" but for some reason this was not working for (also after changing "charging" to "Charging"...).

 infix = variables.ulm_card_battery_battery_state_entity_id !== null
        && states[variables.ulm_card_battery_battery_state_entity_id].state == "charging"
        ? "-charging" : ""