Updated card_media_player.yaml
While adding my Sonos speakers (playing a Spotify playlist), I noticed that both the name and the label were the same. The media_title and media_album_name were exactly the same. This change checks if they are the same, if so, it will show the artist as the label. Otherwise, it will default to the 'on' translation.