Introduce minimum spread variable for thermostat card
Some thermostats don't like when the low and high temperature at set to the same value. The minimum spread will ensure there is always that amount of difference between the `target_temp_low` & `target_temp_high`. It is also able to be set by a user if they'd prefer a higher spread (or a spread of 0).